requirements.txt
kafka-python
streamlit
pandas
numpy
cassandra-driver
